Fix bug caused by blocking in async context
This bug only causes a panic in a debug build since that's the only
place where tokio checks this. It can be reproduced by compiling with
the gst flag.
`cargo build --features gst`
and then running the server as well as the tui.

The problem is caused by the `Instance` type creating a blocking reqwest
Client in `Instance::default`. Since this is eventually called on the creation of a
`Model` type the fix is to change `Model::new` to async and use
`tokio::task::spawn_blocking` to make sure the blocking reqwest client
is created in a context that allows for blocking.

An alternative solution would be to change it so that the reqwest client
is async. However this change would touch a larger part of the codebase
and seems like an unnecessary change at the moment.