Allow "gst" feature for `lib` to allow .opus files
I am not sure why "gst" was uncommented as a feature for `lib` when lib
uses this feature flag to, among other things, condition .opus support.

This commit creates an empty "gst" feature to allow this. I am curently
unable to test without the "gst" flag and as such I decided not to try
to remove all feature flags that use "gst" in the lib.