Large refactor types, daemon-update subscribers
This large refactor commit moves the proto description to lib::types, it
starts placing From and TryFrom implementations in lib::types for gRPC
types.

It also implements a server-to-client gRPC interface where the UI can
subscribe to events from the daemon. This will hopefully make it easier
for the UI to update itself when important events happen in the daemon
    for example when the playing track is changed.