Refactor and finalize DaemonUpdate subscriptions
Finalize the last pieces of the puzzle for implementing the DaemonUpdate
subscription. However implementation of actual daemon update events has
not yet been completed.

In order to listen for daemon updates in a non-blocking or awaiting way
the stream needs to be converted to a channel, this requires an
additional tokio::task.
The updates need to be listened for in a non-blocking manner because
otherwise we block the main ui loop.