Removes ctrl-c from server and dependencies
Control c handler ran in the daemon where it would never be used by the
user. Removes this code and the dependency on ctrl-c.

This functionality does not exist in the TUI currently and adding it
seemed like a lot of work for a feature that we don't currently support.

If we would like to add this to the TUI it would need to be handled like
all other key events since the terminal is running in raw mode. We for
that reason will not have any use for the ctrlc dependency either way.