Refactor Playlist next track logic and cleanup
Remove unecessary function abstractions in the playlist logic. Improve
next track selection by replacing the function with an iterator. This
allows us to not have duplicate songs in the random list (or if we want
the random list to have duplicate songs then that's trivial).

On top of that encode possibility of not having a current track in type
logic. Cleanup and reformat project.