Refactor daemon play-track feature and surrounding
The play-selected-track flow was highly confusing, it involved 3
PlayerCmd's and 1 file write and 1 file read from the TUI and daemon
respectively.

This commit refactors this flow into 1 PlayerCmd and 0 file interaction.
This PR also tries to simplify some other daemon flows by for example
removing Eos signals sent by the daemon to itself when the next track
should be played. Instead this is accomplished by just calling a
function for playing the next track.

The function that plays the next track is also completely reworked and
renamed. It is now called `play_next_track` rather than `start_play`.

This PR also makes sure that the daemon internal state is actually
accurately representing what is going on in the application, previously
the music player may play a track but the `current_track` would not
point to this track.

This commit also gets rid of 2 fields in the GeneralPlayer struct that
turned out to not be necessary after the refactor.

This commit also does some preporatory work for moving towards a model
where playing tracks not in the playlist is possible. For example
playing a track directly from the library.

Finally this commit creates several comments for TODOs and thoughts
about future improvements for the code.