Disable clippy pedantic warnings
We should not have this setting on by default if the codebase is going to be littered with complaints.
Arguably the pedantic settings is too pedantic and gives non-idiomatic
advice.
If we feel like these warnings will serve us in providing better code
then they should be fixed and then the lint should be re-enabled.